Updates Studio service url to a name
When deploying a Juju Application, a Kubernetes service is spawned
bearing the same name as the application. We can use that service
name to resolve any addresses between the services. For example,
we can use http://STUDIO_APP_NAME/studio to access STUDIO, instead of
http://EPHEMERAL_IP:8080/studio.

This way, even if this charm is upgraded or refreshed, we wouldn't
have to update Legend Studio with a new service URL.

Alternatively, we can allow users to bring their own host name
by using the external-hostname config option.